chore: re-organize provider files
